Backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
celery==5.6.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
exceptiongroup==1.3.1
flower==2.0.1
humanize==4.15.0
idna==3.11
kombu==5.6.2
mdb-parser==0.0.3
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==26.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
PyJWT==2.10.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.3.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.6.0
